Added dates section tooltip
