+++ src/js/components/awardv2/visualizations/overview/RelatedAwards.jsx
@@ -5,9 +5,7 @@
 
 import React from 'react';
 import PropTypes from 'prop-types';
-import { throttle } from 'lodash';
 
-import * as Icons from 'components/sharedComponents/icons/Icons';
 import InfoTooltip from '../../idv/InfoTooltip';
 
 const propTypes = {
@@ -15,53 +13,6 @@ const propTypes = {
 };
 
 export default class RelatedAwards extends React.Component {
-    constructor(props) {
-        super(props);
-
-        this.state = {
-            showTooltip: false,
-            offsetTop: 0,
-            offsetRight: 0
-        };
-
-        this.showTooltip = this.showTooltip.bind(this);
-        this.closeTooltip = this.closeTooltip.bind(this);
-        this.measureOffset = throttle(this.measureOffset.bind(this), 16);
-    }
-
-    componentDidMount() {
-        this.measureOffset();
-        window.addEventListener('scroll', this.measureOffset);
-        window.addEventListener('resize', this.measureOffset);
-    }
-
-    componentWillUnmount() {
-        window.removeEventListener('scroll', this.measureOffset);
-        window.removeEventListener('resize', this.measureOffset);
-    }
-
-    showTooltip() {
-        this.setState({
-            showInfoTooltip: true
-        });
-    }
-
-    closeTooltip() {
-        this.setState({
-            showInfoTooltip: false
-        });
-    }
-
-    measureOffset() {
-        const targetElement = this.referenceDiv;
-        const offsetTop = targetElement.offsetTop - 35;
-        const offsetRight = window.innerWidth - targetElement.offsetLeft - targetElement.clientWidth - 300;
-        this.setState({
-            offsetTop,
-            offsetRight
-        });
-    }
-
     render() {
         let parentLink = 'N/A';
         if (this.props.overview.parentAward && this.props.overview.parentId) {
@@ -73,17 +24,11 @@ export default class RelatedAwards extends React.Component {
                 </a>
             );
         }
-        let tooltip = null;
-        if (this.state.showInfoTooltip) {
-            const style = {
-                top: this.state.offsetTop,
-                right: this.state.offsetRight
-            };
-
-            tooltip = (
-                <div
-                    className="info-tooltip-spacer"
-                    style={style}>
+        return (
+            <div
+                className="related-awards">
+                <div className="award-overview__title related-awards__title">
+                    Related Awards
                     <InfoTooltip>
                         <strong>Related Awards</strong>
                         <p>
@@ -98,30 +43,6 @@ export default class RelatedAwards extends React.Component {
                             </li>
                         </ul>
                     </InfoTooltip>
-                </div>
-            );
-        }
-        return (
-            <div
-                className="related-awards">
-                <div className="award-overview__title related-awards__title">
-                    Related Awards
-                    <div className="award__info-wrapper">
-                        <div ref={(div) => {
-                            this.referenceDiv = div;
-                        }}>
-                            <button
-                                onBlur={this.closeTooltip}
-                                className="award__icon"
-                                onFocus={this.showTooltip}
-                                onMouseEnter={this.showTooltip}
-                                onMouseLeave={this.closeTooltip}
-                                onClick={this.showTooltip}>
-                                <Icons.InfoCircle alt="Information" />
-                            </button>
-                            {tooltip}
-                        </div>
-                    </div>
                 </div>
                 <div className="related-awards__parent">
                     <div className="related-awards__label">
